refactor(tests): migrate mobilizations widgets #507
